refactor(api): remove data adapters
